--- internal/usecase/users/forgot_password.go
@@ -1,1 +1,0 @@
-package users
