--- internal/usecase/users/google_login.go
@@ -1,1 +1,0 @@
-package users
