--- internal/usecase/users/update_profile.go
@@ -1,1 +1,0 @@
-package users
