--- internal/usecase/users/verify_email.go
@@ -1,1 +1,0 @@
-package users
