+++ internal/usecase/users_usecase/verify_email.go
@@ -1,0 +1,1 @@
+package users_usecase
